renames main file, removing unused imports
