--- ursa_major/__init__.py
@@ -1,3 +1,0 @@
-# import pymysql
-#
-# pymysql.install_as_MySQLdb()
